Moved the landscape loader and validator to it's own module.
