+++ agents_playground/scene/file/scene.schema.json
@@ -1,0 +1,43 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://agents-playground.com/scene-definition.schema.json",
+  "title": "Scene Definition",
+  "description": "A simulation scene.",
+  "type": "object",
+  "properties": {
+    "file_characteristics": { "$ref": "#/$defs/file_characteristics" }
+  },
+  "required": [],
+  "$defs": {
+    "file_characteristics": {
+      "description": "Metadata about a landscape file.",
+      "type": "object",
+      "properties": {
+        "author":{
+          "description":"Can be any string but intended to be First Name Last Name.",
+          "type": "string"
+        },
+        "license":{
+          "description":"The license type for this attribute.",
+          "type": "string"
+        },
+        "contact":{
+          "description":"Contact information. Could be anything.",
+          "type": "string"
+        },
+        "creation_time":{
+          "description":"When the file was initially created in the format YYYY-MM-DD hh:mm:ss.",
+          "type": "string",
+          "pattern": "[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{6} UTC"
+        },
+        "updated_time":{
+          "description":"When the file was last updated in the format YYYY-MM-DD hh:mm:ss.",
+          "type": "string",
+          "pattern": "[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{6} UTC"
+        }
+      },
+      "required": []
+    }
+  }
+}
+
